revisi format number

+++ database/migrations/2023_01_17_220959_create_purchase_temporary_details_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->string('barcode', 50);
             $table->string('name', 50);
-            $table->double('purchase_price', 15, 4);
-            $table->float('qty');
+            $table->double('purchase_price');
+            $table->double('qty');
             $table->bigInteger('user_id');
             $table->timestamps();
         });
